Replace query with evalQuery
